requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas>=2.0.0
pyarrow>=14.0.0
faker>=24.0.0
boto3>=1.34.0
astronomer-cosmos[dbt-snowflake]>=1.9.0
apache-airflow-providers-snowflake>=5.3.0
apache-airflow-providers-amazon>=8.0.0